+++ java-oop-2/src/Main.java
@@ -5,4 +5,4 @@ public class Main {
         Run run = new Run();
         run.run();
     }
-}햣
+}
