+++ java-oop-2/src/controller/Run.java
@@ -3,8 +3,6 @@ package controller;
 import java.util.Scanner;
 
 import model.Account;
-import model.Deposit;
-import model.Withdraw;
 import view.PrintAccountInfo;
 import view.PrintMessage;
 import view.Menu;
@@ -13,9 +11,6 @@ public class Run {
     Scanner scanner = new Scanner(System.in);
 
     Account account1 = new Account("001", "최정훈", 100000);
-
-    Deposit deposit = new Deposit();
-    Withdraw withdraw = new Withdraw();
 
     PrintAccountInfo printAccountInfo = new PrintAccountInfo();
     PrintMessage printMessage = new PrintMessage();
@@ -36,7 +31,7 @@ public class Run {
                     System.out.print("입금할 금액의 양을 입력하십시오: ");
                     int depositAmount = scanner.nextInt();
 
-                    if (deposit.deposit(account1, depositAmount)) {
+                    if (account1.deposit(account1, depositAmount)) {
                         printMessage.printDeposit(depositAmount, account1.getBalance());
                     } else {
                         printMessage.printDepositError(account1.getBalance());
@@ -47,7 +42,7 @@ public class Run {
                     System.out.print("출금할 금액의 양을 입력하십시오: ");
                     int withdrawAmount = scanner.nextInt();
 
-                    if (withdraw.withdraw(account1, withdrawAmount)) {
+                    if (account1.withdraw(account1, withdrawAmount)) {
                         printMessage.printWithdraw(withdrawAmount, account1.getBalance());
                     } else {
                         printMessage.printWithdrawError(account1.getBalance());
