--- java-oop-2/src/model/Deposit.java
@@ -1,13 +1,0 @@
-package model;
-
-public class Deposit {
-    public boolean deposit(Account account, int amount) {
-        if (amount < 0) {
-            return false;
-        }
-
-        account.setBalance(account.getBalance() + amount);
-
-        return true;
-    }
-}
